backend/requirements.txt
Flask
Flask-Cors
pandas
google-generativeai
python-dotenv
openpyxl
matplotlib